refactor: split chapters logic into button, tooltip, and panel modules for better separation of concerns
